split dashboardMutantFilter and diffMutantFilter
