+++ src/Stryker.Core/Stryker.Core/MutantFilters/DiffMutantFilter.cs
@@ -1,4 +1,4 @@
-﻿using Microsoft.CodeAnalysis;
+using Microsoft.CodeAnalysis;
 using Microsoft.CodeAnalysis.CSharp;
 using Microsoft.CodeAnalysis.Text;
 using Microsoft.Extensions.Logging;
@@ -20,27 +20,17 @@ namespace Stryker.Core.MutantFilters
     public class DiffMutantFilter : IMutantFilter
     {
         private readonly DiffResult _diffResult;
-        private readonly IBaselineProvider _baselineProvider;
-        private readonly IGitInfoProvider _gitInfoProvider;
         private readonly ILogger<DiffMutantFilter> _logger;
 
         private readonly StrykerOptions _options;
-        private readonly JsonReport _baseline;
 
         public string DisplayName => "git diff file filter";
 
-        public DiffMutantFilter(StrykerOptions options, IDiffProvider diffProvider = null, IBaselineProvider baselineProvider = null, IGitInfoProvider gitInfoProvider = null)
+        public DiffMutantFilter(StrykerOptions options, IDiffProvider diffProvider = null)
         {
             _logger = ApplicationLogging.LoggerFactory.CreateLogger<DiffMutantFilter>();
 
             _options = options;
-            _gitInfoProvider = gitInfoProvider ?? new GitInfoProvider(options);
-            _baselineProvider = baselineProvider ?? BaselineProviderFactory.Create(options);
-
-            if (options.CompareToDashboard)
-            {
-                _baseline = GetBaselineAsync().Result;
-            }
 
             _diffResult = diffProvider.ScanDiff();
 
@@ -70,20 +60,6 @@ namespace Stryker.Core.MutantFilters
             // Mutants can be enabled for testing based on multiple reasons. We store all the filtered mutants in this list and return this list.
             IEnumerable<Mutant> filteredMutants;
 
-            // If the dashboard feature is turned on we first filter based on previous results
-            if (options.CompareToDashboard)
-            {
-                // If the dashboard feature is enabled but we cannot find a baseline. We are going to test the entire project. Thus none of the mutants can be filtered out and all are returned.
-                if (_baseline == null)
-                {
-                    _logger.LogDebug("Testing all mutants on {0} because there is no baseline available", file.RelativePathToProjectFile);
-                    return mutants;
-                }
-
-                // Updates all the mutants in this file with their counterpart's result in the report of the previous run
-                UpdateMutantsWithBaselineStatus(mutants, file);
-            }
-
             // A non-csharp file is flagged by the diff result as modified. We cannot determine which mutants will be affected by this, thus all mutants have to be tested.
             if (_diffResult.ChangedTestFiles is { } && _diffResult.ChangedTestFiles.Any(x => !x.EndsWith(".cs")))
             {
@@ -109,112 +85,7 @@ namespace Stryker.Core.MutantFilters
                 filteredMutants = ResetMutantStatusForChangedTests(mutants);
             }
 
-            // Identical mutants within the same file cannot be distinguished from eachother and therefor we cannot give them a mutant status from the baseline. These will have to be re-run.
-            if (_options.CompareToDashboard)
-            {
-                var mutantsNotRun = mutants.Where(x => x.ResultStatus == MutantStatus.NotRun);
-                filteredMutants = MergeMutantLists(filteredMutants, mutantsNotRun);
-            }
-
             return filteredMutants;
-        }
-
-        private void UpdateMutantsWithBaselineStatus(IEnumerable<Mutant> mutants, ReadOnlyFileLeaf file)
-        {
-            var baselineFile = _baseline.Files.SingleOrDefault(f => FilePathUtils.NormalizePathSeparators(f.Key) == FilePathUtils.NormalizePathSeparators(file.RelativePath));
-
-            if (baselineFile is { } && baselineFile.Value is { })
-            {
-                foreach (var baselineMutant in baselineFile.Value.Mutants)
-                {
-                    var baselineMutantSourceCode = GetMutantSourceCode(baselineFile.Value.Source, baselineMutant);
-
-                    if (string.IsNullOrEmpty(baselineMutantSourceCode))
-                    {
-                        _logger.LogWarning("Unable to find mutant span in original baseline source code. This indicates a bug in stryker. Please report this on github.");
-                        continue;
-                    }
-
-                    IEnumerable<Mutant> matchingMutants = GetMutantMatchingSourceCode(mutants, baselineMutant, baselineMutantSourceCode);
-
-                    SetMutantStatusToBaselineMutantStatus(baselineMutant, matchingMutants);
-                }
-            }
-        }
-
-        private static void SetMutantStatusToBaselineMutantStatus(JsonMutant baselineMutant, IEnumerable<Mutant> matchingMutants)
-        {
-            if (matchingMutants.Count() == 1)
-            {
-                matchingMutants.First().ResultStatus = (MutantStatus)Enum.Parse(typeof(MutantStatus), baselineMutant.Status);
-                matchingMutants.First().ResultStatusReason = "Result based on previous run";
-            }
-            else
-            {
-                foreach (var matchingMutant in matchingMutants)
-                {
-                    matchingMutant.ResultStatus = MutantStatus.NotRun;
-                    matchingMutant.ResultStatusReason = "Result based on previous run was inconclusive";
-                }
-            }
-        }
-
-        public string GetMutantSourceCode(string source, JsonMutant baselineMutant)
-        {
-            SyntaxTree tree = CSharpSyntaxTree.ParseText(source);
-
-            var beginLinePosition = new LinePosition(baselineMutant.Location.Start.Line - 1, baselineMutant.Location.Start.Column - 1);
-            var endLinePosition = new LinePosition(baselineMutant.Location.End.Line - 1, baselineMutant.Location.End.Column - 1);
-
-            LinePositionSpan span = new LinePositionSpan(beginLinePosition, endLinePosition);
-
-            var textSpan = tree.GetText().Lines.GetTextSpan(span);
-            var originalNode = tree.GetRoot().DescendantNodes(textSpan).FirstOrDefault(n => textSpan.Equals(n.Span));
-            return originalNode?.ToString();
-
-
-        }
-
-        private IEnumerable<Mutant> GetMutantMatchingSourceCode(IEnumerable<Mutant> mutants, JsonMutant baselineMutant, string baselineMutantSourceCode)
-        {
-            return mutants.Where(x =>
-                x.Mutation.OriginalNode.ToString() == baselineMutantSourceCode &&
-                x.Mutation.DisplayName == baselineMutant.MutatorName);
-        }
-
-        private async Task<JsonReport> GetBaselineAsync()
-        {
-            var branchName = _gitInfoProvider.GetCurrentBranchName();
-
-            var baselineLocation = $"dashboard-compare/{branchName}";
-
-            var report = await _baselineProvider.Load(baselineLocation);
-
-            if (report == null)
-            {
-                _logger.LogInformation("We could not locate a baseline for branch {0}, now trying fallback version {1}", branchName, _options.FallbackVersion);
-
-                return await GetFallbackBaselineAsync();
-            }
-
-            _logger.LogInformation("Found baseline report for current branch {0}", branchName);
-
-            return report;
-        }
-
-        private async Task<JsonReport> GetFallbackBaselineAsync()
-        {
-            var report = await _baselineProvider.Load(_options.FallbackVersion);
-
-            if (report == null)
-            {
-                _logger.LogInformation("We could not locate a baseline report for the current branch or fallback version. Now running a complete test to establish a baseline.");
-                return null;
-            }
-
-            _logger.LogInformation("Found fallback report using version {0}", _options.FallbackVersion);
-
-            return report;
         }
 
         private IEnumerable<Mutant> SetNotRunMutantsToIgnored(IEnumerable<Mutant> mutants)
@@ -270,35 +141,6 @@ namespace Stryker.Core.MutantFilters
             }
 
             return filteredMutants;
-        }
-
-        /// Takes two lists. Adds the mutants from the updateMutants list to the targetMutants.
-        /// If the targetMutants already contain a member with the same Id. The results are updated.
-        internal IEnumerable<Mutant> MergeMutantLists(IEnumerable<Mutant> target, IEnumerable<Mutant> source)
-        {
-            foreach (var targetMutant in target)
-            {
-                if (source.Any(updateMutant => updateMutant.Id == targetMutant.Id))
-                {
-                    continue;
-                }
-
-                yield return targetMutant;
-            }
-
-            foreach (var sourceMutant in source)
-            {
-                if (target.SingleOrDefault(targetMutant => targetMutant.Id == sourceMutant.Id) is var targetMutant && targetMutant is { })
-                {
-                    targetMutant.ResultStatus = sourceMutant.ResultStatus;
-                    targetMutant.ResultStatusReason = sourceMutant.ResultStatusReason;
-
-                    yield return targetMutant;
-                    continue;
-                }
-
-                yield return sourceMutant;
-            }
         }
     }
 }
