+++ src/Stryker.Core/Stryker.Core/MutantFilters/MutantFilterFactory.cs
@@ -1,4 +1,4 @@
-﻿using Stryker.Core.Baseline;
+using Stryker.Core.Baseline;
 using Stryker.Core.DashboardCompare;
 using Stryker.Core.DiffProviders;
 using Stryker.Core.Options;
@@ -36,9 +36,13 @@ namespace Stryker.Core.MutantFilters
                     new ExcludeFromCodeCoverageFilter()
                 };
 
+            if (options.CompareToDashboard)
+            {
+                enabledFilters.Add(new DashboardMutantFilter(options, _baselineProvider, _gitInfoProvider));
+            }
             if (options.DiffEnabled)
             {
-                enabledFilters.Add(new DiffMutantFilter(options, _diffProvider, baselineProvider: _baselineProvider, gitInfoProvider: _gitInfoProvider));
+                enabledFilters.Add(new DiffMutantFilter(options, _diffProvider));
             }
 
             return enabledFilters;
